starting over. parsing combinators tried for deserializing
unfortunately, the binary format itself isn't self-explanatory, a final parser is type-specific and is to be created for a every certain type.
also, parsing combinators are expected to be applied for serialization, for which purpose an intermediate model introduced, and an attempt to build it using computation expressions made.
abstract from real model, just a concept.